[3d] refactor code
